requirements.txt
streamlit == 1.37.0
langchain == 0.2.12
langchain-openai == 0.1.20
langchain-community == 0.2.11
python-dotenv == 1.0.1
beautifulsoup4 == 4.12.3
faiss-cpu == 1.8.0.post1
pdfminer.six == 20240706